requirements.txt
beautifulsoup4==4.11.1
lorem==0.1.1
matplotlib==3.5.1
pandas==1.3.2
Pillow==9.2.0
requests==2.26.0
streamlit==1.12.2
wordcloud==1.8.1
